changed to using auth hook for contexts
