+++ bookkeeping-app/src/components/contexts/AuthCtx.jsx
@@ -1,158 +1,10 @@
-import { createContext, useState, useEffect } from "react";
+import { createContext } from "react";
 
-import { BASE_URL } from "../../constants";
-import { useToast } from "./ToastCtx";
-
-const AuthCtx = createContext({
-    ctxAccessToken: null,
-    setCtxAccessToken: () => {},
-    ctxUserData: null,
-    ctxUpdateUser: () => {},
-    ctxUpdatePwd: () => {},
-    requestPswdReset: () => {},
-    logoutUser: () => {},
-});
+const AuthCtx = createContext({});
 
 export function AuthCtxProvider(props) {
-    const { showToast } = useToast();
-
-    const [ctxAccessToken, setCtxAccessToken] = useState(localStorage.getItem("accessToken") || null);
-    const [ctxUserData, setCtxUserData] = useState({});
-
-    const baseUrl = BASE_URL;
-
-    useEffect(() => {
-        if (ctxAccessToken) {
-            ctxGetUser();
-        }
-    }, [ctxAccessToken]);
-
-    const ctxGetUser = async () => {
-        try {
-            const response = await fetch(`${baseUrl}/api/profile/`, {
-                method: "GET",
-                headers: {
-                    "Content-Type": "application/json",
-                    Authorization: `Bearer ${ctxAccessToken}`,
-                },
-            });
-
-            if (!response.ok) {
-                throw new Error(`HTTP error! status: ${response.status}`);
-            } else {
-                const returnedProfile = await response.json();
-                setCtxUserData(returnedProfile);
-            }
-        } catch (e) {
-            console.log("Error: " + e);
-        }
-    };
-
-    const ctxUpdateUser = async (updatedUser) => {
-        console.log(updatedUser);
-        try {
-            const response = await fetch(`${baseUrl}/api/profile/`, {
-                method: "PUT",
-                headers: {
-                    "Content-Type": "application/json",
-                    Authorization: `Bearer ${ctxAccessToken}`,
-                },
-                body: JSON.stringify(updatedUser),
-            });
-
-            if (!response.ok) {
-                throw new Error(`HTTP error! status: ${response.status}`);
-            } else {
-                const returnedProfile = await response.json();
-                console.log(returnedProfile);
-                showToast("Profile updated", "success", 3000);
-            }
-        } catch (e) {
-            console.log("Error: " + e);
-            showToast("Error updating Profile", "error", 5000);
-        }
-    };
-
-    const ctxUpdatePwd = async (pwdCurr, pwdNew, pwdCnfm) => {
-        try {
-            const response = await fetch(`${baseUrl}/api/auth/users/set_password/`, {
-                method: "POST",
-                headers: {
-                    "Content-Type": "application/json",
-                    Authorization: `Bearer ${ctxAccessToken}`,
-                },
-                body: JSON.stringify({
-                    current_password: pwdCurr,
-                    new_password: pwdNew,
-                    re_new_password: pwdCnfm,
-                }),
-            });
-
-            if (!response.ok) {
-                const errorData = await response.json();
-                let errorMessage = "Unknown error";
-
-                if (errorData.current_password) {
-                    errorMessage = errorData.current_password[0];
-                } else if (errorData.new_password) {
-                    errorMessage = errorData.new_password[0];
-                } else if (errorData.re_new_password) {
-                    errorMessage = errorData.re_new_password[0];
-                } else if (errorData.detail) {
-                    errorMessage = errorData.detail;
-                }
-
-                showToast("Error updating profile", "error", 5000);
-                console.log(`Password update failed. Status: ${response.status}. Error: ${errorMessage}`);
-                return errorMessage;
-            } else {
-                showToast("Password updated", "success", 3000);
-                return "";
-            }
-        } catch (e) {
-            console.log("Error: " + e);
-            showToast("Network error updating Password", "error", 5000);
-        }
-    };
-
-    const requestPswdReset = async (email) => {
-        try {
-            const response = await fetch(`${baseUrl}/api/auth/users/reset_password/`, {
-                method: "POST",
-                headers: {
-                    "Content-Type": "application/json",
-                },
-                body: JSON.stringify({ email: email }),
-            });
-
-            if (response.status === 204 || response.ok) {
-                return "Success! If the email is registered, you will receive a reset link shortly.";
-            } else {
-                console.error("Reset Password Request Error Status:", response.status);
-                return "An error occurred. Please try again.";
-            }
-        } catch (e) {
-            console.error("Network Error:", e);
-            return "A network error occurred. Please check your connection.";
-        }
-    };
-
-    const logoutUser = () => {
-        localStorage.removeItem("accessToken");
-        setCtxAccessToken(null);
-    };
-
-    const context = {
-        ctxAccessToken,
-        setCtxAccessToken,
-        ctxUserData,
-        ctxUpdateUser,
-        ctxUpdatePwd,
-        requestPswdReset,
-        logoutUser,
-    };
-
-    return <AuthCtx.Provider value={context}>{props.children}</AuthCtx.Provider>;
+    // Only keep context-specific values here, e.g. for global state or cross-component sharing
+    return <AuthCtx.Provider value={{}}>{props.children}</AuthCtx.Provider>;
 }
 
 export default AuthCtx;
