+++ bookkeeping-app/src/components/contexts/EntitiesCtx.jsx
@@ -3,8 +3,8 @@ import useSWRImmutable from "swr/immutable";
 
 import { BASE_URL } from "../../constants";
 import { useToast } from "./ToastCtx";
-import AuthCtx from "./AuthCtx";
 import PropertiesCtx from "./PropertiesCtx";
+import { UseAuth } from "../../hooks/UseAuth";
 
 const EntitiesCtx = createContext({
     ctxEntityList: null,
@@ -19,7 +19,7 @@ const EntitiesCtx = createContext({
 export function EntitiesCtxProvider(props) {
     const { showToast } = useToast();
 
-    const { ctxAccessToken } = useContext(AuthCtx);
+    const { accessToken: ctxAccessToken } = UseAuth();
     const { ctxActiveProperty } = useContext(PropertiesCtx);
 
     const [ctxActiveEntity, setCtxActiveEntity] = useState();
