+++ bookkeeping-app/src/components/contexts/PropertiesCtx.jsx
@@ -3,7 +3,7 @@ import useSWRImmutable from "swr/immutable";
 
 import { BASE_URL } from "../../constants";
 import { useToast } from "./ToastCtx";
-import AuthCtx from "./AuthCtx";
+import { UseAuth } from "../../hooks/UseAuth";
 
 const PropertiesCtx = createContext({
     ctxActiveProperty: null,
@@ -18,7 +18,7 @@ const PropertiesCtx = createContext({
 export function PropertiesCtxProvider(props) {
     const { showToast } = useToast();
 
-    const { ctxAccessToken } = useContext(AuthCtx);
+    const { accessToken: ctxAccessToken } = UseAuth();
 
     const [ctxActiveProperty, setCtxActiveProperty] = useState();
