+++ bookkeeping-app/src/components/contexts/RentPaymentsCtx.jsx
@@ -3,8 +3,8 @@ import useSWRImmutable from "swr/immutable";
 
 import { BASE_URL } from "../../constants";
 import { useToast } from "./ToastCtx";
-import AuthCtx from "./AuthCtx";
 import PropertiesCtx from "./PropertiesCtx";
+import { UseAuth } from "../../hooks/UseAuth";
 
 const RentPaymentsCtx = createContext({
     ctxMonthPaymentList: null,
@@ -26,7 +26,7 @@ const getInitialDate = () => {
 export function RentPaymentsCtxProvider(props) {
     const { showToast } = useToast();
 
-    const { ctxAccessToken } = useContext(AuthCtx);
+    const { accessToken: ctxAccessToken } = UseAuth();
     const { ctxActiveProperty } = useContext(PropertiesCtx);
 
     const [ctxActiveDate, setCtxActiveDate] = useState(getInitialDate());
